Moved caching related classes to under validator.caching package.
