+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/validator/caching/CachingDamlLedgerStateReader.scala
@@ -1,11 +1,17 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.validator
+package com.daml.ledger.validator.caching
 
 import com.daml.caching.Cache
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.{DamlStateKey, DamlStateValue}
 import com.daml.ledger.validator.LedgerStateOperations.Key
+import com.daml.ledger.validator.{
+  DamlLedgerStateReader,
+  LedgerStateReader,
+  RawToDamlLedgerStateReaderAdapter,
+  StateKeySerializationStrategy
+}
 
 import scala.collection.mutable
 import scala.concurrent.{ExecutionContext, Future}
